requirements.txt
blinker==1.7.0
click==8.1.7
filelock==3.13.1
flask==3.0.2
fsspec==2024.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
kaleido==0.2.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pillow==10.2.0
plotly==5.19.0
sympy==1.12
tenacity==8.2.3
torch==2.2.1
torchvision==0.17.1
triton==2.2.0
typing-extensions==4.10.0
werkzeug==3.0.1
zipp==3.17.0
